changed tag_add to add_tag and moved the function is_entry_tagged to keepass.py
